consolidates all auth styling into one file, discards unsused style files
