+++ components/SignUp.js
@@ -10,7 +10,7 @@ import {
 } from "react-native";
 import { useDispatch, useSelector } from "react-redux";
 import { registerUser, clearError } from "../store/auth/authActions";
-import styles from "../styles/signUpStyles.js";
+import styles from "../styles/authPageStyles";
 import RecipeShareLogo from "./RecipeShareLogo.js";
 import { maxUsername } from "../constants/maxLenth";
 
@@ -49,7 +49,7 @@ const SignUp = ({ navigation }) => {
                 <Text style={styles.explanationText}>
                     Sign up to start editing your favorite recipes.
                 </Text>
-                <Text style={styles.emailText}>Username</Text>
+                <Text style={styles.inputText}>Username</Text>
                 <TextInput
                     ref={usernameInput}
                     style={styles.inputFields}
@@ -65,7 +65,7 @@ const SignUp = ({ navigation }) => {
                     style={styles.maxLengthIndicator}
                 >{`${signUp.username.length}/${maxUsername}`}</Text>
 
-                <Text style={styles.passwordText}>Password</Text>
+                <Text style={styles.inputText}>Password</Text>
                 <TextInput
                     ref={passwordInput}
                     style={styles.inputFields}
@@ -84,7 +84,7 @@ const SignUp = ({ navigation }) => {
                         navigation.navigate("Login");
                     }}
                 >
-                    <Text style={styles.loginButton}>
+                    <Text style={styles.switchAuthPageLink}>
                         Have an account? Login
                     </Text>
                 </TouchableOpacity>
@@ -92,14 +92,12 @@ const SignUp = ({ navigation }) => {
                 <View style={{ flexDirection: "row-reverse" }}>
                     <TouchableOpacity
                         onPress={register}
-                        style={styles.createAccountButton}
+                        style={styles.submitButton}
                     >
                         {isLoading ? (
                             <ActivityIndicator size="large" color="#00ff00" />
                         ) : (
-                            <Text style={styles.createAccountText}>
-                                Sign Up
-                            </Text>
+                            <Text style={styles.submitButtonText}>Sign Up</Text>
                         )}
                     </TouchableOpacity>
                 </View>
