+++ styles/authPageStyles.js
@@ -3,6 +3,7 @@ import { StyleSheet } from "react-native";
 const styles = StyleSheet.create({
     container: {
         paddingHorizontal: 16,
+        paddingVertical: 20,
     },
     title: {
         textAlign: "center",
@@ -11,16 +12,10 @@ const styles = StyleSheet.create({
         fontWeight: "bold",
     },
     explanationText: {
-        textAlign: "left",
+        textAlign: "center",
         color: "#363838",
-
         fontSize: 16,
         marginVertical: 20,
-    },
-    createAccountButton: {
-        fontSize: 14,
-        fontWeight: "500",
-        color: "#047396",
     },
     inputFields: {
         height: 40,
@@ -31,26 +26,32 @@ const styles = StyleSheet.create({
         borderWidth: 1,
         borderColor: "#d6d7da",
     },
-    loginButton: {
+    inputText: {
+        color: "#363838",
+    },
+    submitButton: {
         justifyContent: "center",
-        width: 120,
+        width: 217,
         height: 40,
         borderRadius: 50,
         backgroundColor: "#3BA405",
         paddingVertical: 10,
         marginTop: 30,
     },
-    emailText: {
-        color: "#363838",
-    },
-    passwordText: {
-        color: "#363838",
-    },
-    loginButtonText: {
+    submitButtonText: {
         textAlign: "center",
         color: "white",
         fontSize: 16,
         fontWeight: "bold",
+    },
+    switchAuthPageLink: {
+        fontSize: 14,
+        fontWeight: "500",
+        color: "#047396",
+    },
+    maxLengthIndicator: {
+        alignSelf: "flex-end",
+        fontSize: 12,
     },
 });
